# prpo-vaje/Klik-kafe-rest/src/main/resources/doc/v1/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentation for NarociloREST, KavarnaREST and UporabnikREST",
    "version" : "1.0.0",
    "title" : "Documentation for Klik-kafe API",
    "contact" : {
      "email" : "info@projekti.prpo.si"
    }
  },
  "host" : "localhost",
  "basePath" : "/v1",
  "tags" : [ {
    "name" : "kavarne"
  }, {
    "name" : "narocila"
  }, {
    "name" : "uporabniki"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/kavarne" : {
      "get" : {
        "tags" : [ "kavarne" ],
        "summary" : "return all kavarnas",
        "description" : "return list of all cavarnas in database",
        "operationId" : "returnAll",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Kavarna"
              }
            }
          }
        }
      }
    },
    "/kavarne/{id}" : {
      "get" : {
        "tags" : [ "kavarne" ],
        "summary" : "return kavarna",
        "description" : "return kavarna by chosen ID",
        "operationId" : "returnById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Kavarna"
            }
          }
        }
      }
    },
    "/narocila" : {
      "get" : {
        "tags" : [ "narocila" ],
        "summary" : "get all orders",
        "description" : "orders are return in pages of size specified in step",
        "operationId" : "getOrders",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 0,
          "format" : "int32"
        }, {
          "name" : "step",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 10,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Narocilo"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "narocila" ],
        "summary" : "submit new order",
        "description" : "submit new order for user with userID",
        "operationId" : "submitOrder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "uporabnik",
          "in" : "header",
          "required" : false,
          "type" : "integer",
          "default" : "",
          "format" : "int32"
        }, {
          "name" : "kavarna",
          "in" : "header",
          "required" : false,
          "type" : "integer",
          "default" : "",
          "format" : "int32"
        }, {
          "name" : "size",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "napitek",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "items" : {
            "type" : "string"
          },
          "default" : ""
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/narocila/{id}" : {
      "get" : {
        "tags" : [ "narocila" ],
        "summary" : "get order for chosen ID",
        "description" : "return order with chosen ID",
        "operationId" : "getOrder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Narocilo"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "narocila" ],
        "summary" : "delete order",
        "description" : "delete order with chosen ID",
        "operationId" : "cancelOrder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/narocila/{userId}/uporabnik" : {
      "get" : {
        "tags" : [ "narocila" ],
        "summary" : "get order for chosen user ID",
        "description" : "All orders from user with this ID",
        "operationId" : "getUserOrders",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Narocilo"
            }
          }
        }
      }
    },
    "/uporabniki" : {
      "get" : {
        "tags" : [ "uporabniki" ],
        "summary" : "get all users",
        "description" : "get information for all users in database",
        "operationId" : "getUsers",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Uporabnik"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "uporabniki" ],
        "summary" : "add new user",
        "description" : "add new user",
        "operationId" : "addUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "surname",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "username",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "name",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "latitude",
          "in" : "header",
          "required" : false,
          "type" : "number",
          "default" : "",
          "format" : "double"
        }, {
          "name" : "longitude",
          "in" : "header",
          "required" : false,
          "type" : "number",
          "default" : "",
          "format" : "double"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/uporabniki/{id}" : {
      "get" : {
        "tags" : [ "uporabniki" ],
        "summary" : "get user",
        "description" : "get user information by chosen ID",
        "operationId" : "getUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json", "text/plain" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Uporabnik"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "uporabniki" ],
        "summary" : "update user",
        "description" : "update user information with chosen ID",
        "operationId" : "updateUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "uporabniki" ],
        "summary" : "delete user",
        "description" : "delete user with chosen ID",
        "operationId" : "deleteUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "Kavarna" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "latitude" : {
          "type" : "number",
          "format" : "double"
        },
        "longitude" : {
          "type" : "number",
          "format" : "double"
        },
        "name" : {
          "type" : "string"
        },
        "napiteks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Napitek"
          }
        },
        "narocilos" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Narocilo"
          }
        }
      }
    },
    "Napitek" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "prepTime" : {
          "type" : "integer",
          "format" : "int32"
        },
        "price" : {
          "type" : "number",
          "format" : "double"
        },
        "size" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "kavarnas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Kavarna"
          }
        },
        "napitkiNarocilas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Napitki_narocila"
          }
        }
      }
    },
    "Napitki_narocila" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "$ref" : "#/definitions/Napitki_narocilaPK"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "napitek" : {
          "$ref" : "#/definitions/Napitek"
        },
        "narocilo" : {
          "$ref" : "#/definitions/Narocilo"
        }
      }
    },
    "Napitki_narocilaPK" : {
      "type" : "object",
      "properties" : {
        "idNarocila" : {
          "type" : "integer",
          "format" : "int32"
        },
        "idNapitka" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Narocilo" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "paymentStatus" : {
          "type" : "string"
        },
        "prepStatus" : {
          "type" : "string"
        },
        "prepTime" : {
          "type" : "integer",
          "format" : "int32"
        },
        "totalPrice" : {
          "type" : "number",
          "format" : "double"
        },
        "napitkiNarocilas" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Napitki_narocila"
          }
        },
        "kavarna" : {
          "$ref" : "#/definitions/Kavarna"
        },
        "uporabnik" : {
          "$ref" : "#/definitions/Uporabnik"
        }
      }
    },
    "Uporabnik" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "email" : {
          "type" : "string"
        },
        "latitude" : {
          "type" : "number",
          "format" : "double"
        },
        "longitude" : {
          "type" : "number",
          "format" : "double"
        },
        "name" : {
          "type" : "string"
        },
        "surname" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "narocilos" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Narocilo"
          }
        }
      }
    }
  }
}